Well, sorted out the OBJECT_MODE gendering bug.
Turns out that it doesn't make heaps of sense to represent these using the same logic as the other stuff. There were a couple of hints along the way:

- The rendering algorithm is totally different
- They're not listed as actual unicode characters, but instead as unicode ZWJ sequences.

Might be worth subclassing or making them all implement a common interface or something. I'll probably have more ideas about this if/when I add families, which are _also_ ZWJ sequences.